requirements.txt
requests
json
python-dotenv
beautifulsoup4
streamlit
openai